[Shared] Move type Name from backend to shared
