+++ services/backend/src/models/course.ts
@@ -14,7 +14,7 @@ import {
   UpdatedAt,
 } from 'sequelize-typescript'
 
-import { Name } from '../types'
+import { Name } from '../shared/types'
 import { CourseProvider } from './courseProvider'
 import { CourseType } from './courseType'
 import { Credit } from './credit'
