+++ services/backend/src/models/semester.ts
@@ -1,8 +1,7 @@
-/* eslint-disable import/no-cycle */
 import { InferAttributes } from 'sequelize'
 import { Column, CreatedAt, DataType, Model, PrimaryKey, Table, UpdatedAt } from 'sequelize-typescript'
 
-import { Name } from '../types'
+import { Name } from '../shared/types'
 
 @Table({
   underscored: true,
