+++ services/backend/src/services/courses/index.ts
@@ -3,7 +3,8 @@ import { Op } from 'sequelize'
 
 import { serviceProvider } from '../../config'
 import { Course, Credit, Enrollment, Organization, SISStudyRightElement } from '../../models'
-import { EnrollmentState, Name, Unification } from '../../types'
+import { Name } from '../../shared/types'
+import { EnrollmentState, Unification } from '../../types'
 import { isOpenUniCourseCode } from '../../util'
 import logger from '../../util/logger'
 import { getSortRank } from '../../util/sortRank'
