+++ services/backend/src/services/faculty/facultyCredits.ts
@@ -1,4 +1,4 @@
-import { Name } from '../../types'
+import { Name } from '../../shared/types'
 import { getCreditsProduced } from '../providerCredits'
 import { isRelevantProgramme } from '../studyProgramme/studyProgrammeHelpers'
 import { getDegreeProgrammesOfFaculty } from './faculty'
