+++ services/backend/src/services/populations/closeToGraduation.ts
@@ -1,8 +1,9 @@
 import { col, Op } from 'sequelize'
 
 import { Course, Credit, Organization, Student, Studyplan, SISStudyRight, SISStudyRightElement } from '../../models'
+import { Name } from '../../shared/types'
 import { mapToProviders } from '../../shared/util'
-import { CreditTypeCode, DegreeProgrammeType, ExtentCode, Name, SemesterEnrollment } from '../../types'
+import { CreditTypeCode, DegreeProgrammeType, ExtentCode, SemesterEnrollment } from '../../types'
 import { redisClient } from '../redis'
 import { getCurriculumVersion } from './shared'
