+++ services/backend/src/services/populations/shared.ts
@@ -6,7 +6,8 @@ import { Op, QueryTypes } from 'sequelize'
 import { dbConnections } from '../../database/connection'
 import { Course, Credit, Enrollment, SISStudyRight, SISStudyRightElement, Student, Studyplan } from '../../models'
 import { Tag, TagStudent } from '../../models/kone'
-import { Criteria, DegreeProgrammeType, EnrollmentState, Name, ParsedCourse } from '../../types'
+import { Name } from '../../shared/types'
+import { Criteria, DegreeProgrammeType, EnrollmentState, ParsedCourse } from '../../types'
 import { SemesterStart } from '../../util/semester'
 import { hasTransferredFromOrToProgramme } from '../studyProgramme/studyProgrammeHelpers'
