+++ services/backend/src/services/studyProgramme/studyProgrammeCourses.ts
@@ -2,8 +2,9 @@ import { orderBy, range } from 'lodash'
 import { Op } from 'sequelize'
 
 import { Credit, Enrollment } from '../../models'
+import { Name } from '../../shared/types'
 import { mapToProviders } from '../../shared/util'
-import { CreditTypeCode, EnrollmentState, Name } from '../../types'
+import { CreditTypeCode, EnrollmentState } from '../../types'
 import { isOpenUniCourseCode } from '../../util'
 import { createArrayOfCourses } from '../languageCenterData'
 import {
