+++ services/backend/src/services/teachers/index.ts
@@ -2,7 +2,8 @@ import { Op, QueryTypes } from 'sequelize'
 
 import { dbConnections } from '../../database/connection'
 import { Credit, Course, Semester, Teacher } from '../../models'
-import { CreditTypeCode, Name } from '../../types'
+import { Name } from '../../shared/types'
+import { CreditTypeCode } from '../../types'
 import { splitByEmptySpace } from '../../util'
 import { isRegularCourse, TeacherStats } from './helpers'
