+++ services/backend/src/types/index.ts
@@ -11,7 +11,6 @@ export { Graduated, SpecialGroups, StatsType, ProgrammeFilter, YearType } from '
 export { GenderCode } from './genderCode'
 export { IamAccess } from './iamAccess'
 export { Language } from './language'
-export { Name } from './name'
 export { Phase } from './phase'
 export { Role } from './role'
 export { SemesterEnrollment } from './semesterEnrollment'
