+++ services/shared/types/index.ts
@@ -1,0 +1,1 @@
+export { Name } from './name'
